Update VenuesPage.vue (#23)
* Update VenuesPage.vue

I wrote all the logic of the filters and copy the style of the accomodations

* hello

* Venue.js and VenueList.vue

I improved the code on venue.js to have more personalised venues and I did the details pop-up. It is not finished yet, it still needing an style improve. And the function confirmationButton needs to be done.

* VenuePage.vue

I styled the VenuePage.vue and the VenueList.vue, randomized the values of the Restaurant etc... and i created a price and rating order button. It miss the confirmation button but it is finished.

* Title on VenuePage and 1 missing error on VenueList